Domains

Types

ClimaCore.Domains.AbstractDomain — Type

AbstractDomain

A domain represents a region of space.

ClimaCore.Domains.IntervalDomain — Type

IntervalDomain(coord⁻, coord⁺; periodic=true)
IntervalDomain(coord⁻, coord⁺; boundary_names::Tuple{Symbol,Symbol})

Construct a IntervalDomain, the closed interval is given by coord⁻, coord⁺ coordinate arguments.

Either a periodic or boundary_names keyword argument is required.

ClimaCore.Domains.RectangleDomain — Type

RectangleDomain(x1::ClosedInterval, x2::ClosedInterval;
    x1boundary::Tuple{Symbol,Symbol},
    x2boundary::Tuple{Symbol,Symbol},
    x1periodic = false,
    x2periodic = false,
)

Construct a RectangularDomain in the horizontal. If a given x1 or x2 boundary is not periodic, then x1boundary or x2boundary boundary name keyword arguments must be supplied.

ClimaCore.Domains.SphereDomain — Type

SphereDomain(radius)

A domain representing the surface of a sphere with radius radius.

Interfaces

ClimaCore.Domains.boundary_names — Function

boundary_names(obj::Union{AbstractDomain, AbstractMesh, AbstractTopology})

A tuple or vector of unique boundary names of a spatial domain.