Grids

ClimaCore.Grids.CellFace — Type

CellFace()

Cell face location

ClimaCore.Grids.CellCenter — Type

CellCenter()

Cell center location

ClimaCore.Grids.ColumnGrid — Type

ColumnGrid(
    full_grid :: ExtrudedFiniteDifferenceGrid, 
    colidx    :: ColumnIndex,
)

A view into a column of a ExtrudedFiniteDifferenceGrid. This can be used as an

ClimaCore.Grids.FiniteDifferenceGrid — Type

FiniteDifferenceGrid(topology::Topologies.IntervalTopology)
FiniteDifferenceGrid(device::ClimaComms.AbstractDevice, mesh::Meshes.IntervalMesh)

Construct a FiniteDifferenceGrid from an IntervalTopology (or an IntervalMesh).

This is an object which contains all the necessary geometric information.

To avoid unnecessary duplication, we memoize the construction of the grid.

ClimaCore.Grids.ExtrudedFiniteDifferenceGrid — Type

ExtrudedFiniteDifferenceGrid(
    horizontal_space::AbstractSpace,
    vertical_space::FiniteDifferenceSpace,
    hypsography::HypsographyAdaption = Flat(),
)

Construct an ExtrudedFiniteDifferenceGrid from the horizontal and vertical spaces.

ClimaCore.Grids.SpectralElementGrid1D — Type

SpectralElementGrid1D(mesh::Meshes.IntervalMesh, quadrature_style::Quadratures.QuadratureStyle)

A one-dimensional space: within each element the space is represented as a polynomial.

ClimaCore.Grids.SpectralElementGrid2D — Type

SpectralElementSpace2D <: AbstractSpace

A two-dimensional space: within each element the space is represented as a polynomial.

Hypsography

ClimaCore.Grids.Flat — Type

Flat()

No surface hypsography.