API

DataLayouts

ClimateMachineCore.DataLayouts — Module

ClimateMachineCore.DataLayouts

Notation:

i,j are horizontal node indices within an element

k is the vertical node index within an element

v is the vertical element index in a stack

h is the element stack index

f is the field index

Data layout is specified by the order in which they appear, e.g. IJKFVH indexes the underlying array as [i,j,k,f,v,h]

Topologies

ClimateMachineCore.Topologies — Module

ClimateMachineCore.Topologies

Objects describing the horizontal connections between elements.

All elements are quadrilaterals, using the face and vertex numbering convention from p4est:

          4
      3-------4
 ^    |       |
 |  1 |       | 2
x2    |       |
      1-------2
          3
        x1-->

Meshes

ClimateMachineCore.Meshes — Module

Meshes

domain

topology

coordinates

metric terms (inverse partial derivatives)

quadrature rules and weights

https://ceed.exascaleproject.org/ceed-code/

QA: https://github.com/CliMA/ClimateMachine.jl/blob/ans/sphere/test/Numerics/DGMethods/compressiblenavierstokesequations/sphere/spherehelper_functions.jl

ClimateMachineCore.Meshes.Quadratures.degrees_of_freedom — Function

degrees_of_freedom(QuadratureStyle) -> Int

Returns the degreesoffreedom of the QuadratureStyle concrete type

ClimateMachineCore.Meshes.Quadratures.GL — Type

GL{Nq}()

Gauss-Legendre quadrature using Nq quadrature points.

ClimateMachineCore.Meshes.Quadratures.quadrature_points — Function

points, weights = quadrature_points(::Type{FT}, quadrature_style)

The points and weights of the quadrature rule in floating point type FT.

ClimateMachineCore.Meshes.Quadratures.GLL — Type

GLL{Nq}()

Gauss-Legendre-Lobatto quadrature using Nq quadrature points.

ClimateMachineCore.Meshes.Quadratures.polynomial_degree — Function

polynomial_degree(QuadratureStyle) -> Int

Returns the polynomial degree of the QuadratureStyle concrete type

ClimateMachineCore.Meshes.Quadratures.QuadratureStyle — Type

QuadratureStyle