ClimaCoreTempestRemap.jl

ClimaCoreTempestRemap.jl provides an interfaces for using ClimaCore meshes with the TempestRemap remapping package, by Paul Ullrich.

Interface

ClimaCoreTempestRemap.write_exodus — Function

write_exodus(filename, topology::Topology2D; normalize_coordinates=true)

Write the topology to an Exodus-formatted NetCDF file.

It tries to adhere to the Exodus II specification, but it is primarily intended for use with TempestRemap.

Note: the generated meshes will use a different ordering of nodes and elements than those generated by TempestRemap itself.

Options:

normalize_coordinates: if true, the coordinates are normalized to be on the unit sphere (this is required for use with TempestRemap)

References

EXODUS II: A finite element data model: https://www.osti.gov/biblio/10102115-exodus-ii-finite-element-data-model