DSS

Direct stiffness summation makes a field on a continuous spectral-element space single-valued at element boundaries (DSS and numerical fluxes). Spaces.weighted_dss! is the user-level entry point; the Topologies functions are the phases it is built from.

ClimaCore.Topologies.dss_transform — Function

dss_transform(arg, local_geometry, weight, I)

Transform arg[I] to a basis for direct stiffness summation (DSS) and multiply it by the DSS weight. Transformations only apply to vector quantities.

local_geometry[I]: The LocalGeometry at I. If local_geometry is nothing, no transformation is performed.

weight[I]: The DSS weight at I. If weight is nothing, the weight is 1 and the result is plain summation.

See ClimaCore.Spaces.weighted_dss!.

ClimaCore.Topologies.dss_transform! — Function

dss_transform!(device, dss_buffer, data, local_geometry, dss_weights, perimeter, localelems)

Transform vectors in data from covariant/contravariant axes to physical axes, weight the data at perimeter nodes, and store the result in dss_buffer.perimeter_data.

Arguments

dss_buffer: DSSBuffer created by create_dss_buffer for data.

data: Field data.

local_geometry: Local metric information at each node.

dss_weights: DSS weights of the horizontal space.

perimeter: Perimeter iterator.

localelems: Local elements on which to perform the transformation.

Part of ClimaCore.Spaces.weighted_dss!.

ClimaCore.Topologies.dss_untransform! — Function

dss_untransform!(device, dss_buffer, data, local_geometry, perimeter, localelems)

Transform physical vectors in dss_buffer.perimeter_data back to their original covariant/contravariant axes, and store the result in data.

Arguments

dss_buffer: DSSBuffer created by create_dss_buffer for data.

data: Field data.

local_geometry: Local metric information at each node.

perimeter: Perimeter iterator.

localelems: Local elements on which to perform the transformation.

Part of ClimaCore.Spaces.weighted_dss!.

ClimaCore.Topologies.dss_untransform — Function

dss_untransform(T, targ, local_geometry, I)

Transform targ back to a value of type T using local_geometry[I], after direct stiffness summation (DSS).

See ClimaCore.Spaces.weighted_dss!.

ClimaCore.Topologies.dss_local! — Function

dss_local!(device, perimeter_data, perimeter, topology)

Perform DSS on the local vertices and interior faces of topology: the values in perimeter_data at nodes shared between local elements are replaced by their sum.

Part of ClimaCore.Spaces.weighted_dss!.

ClimaCore.Topologies.dss_local_ghost! — Function

dss_local_ghost!(device, perimeter_data, perimeter, topology)

Compute the "local" part of the ghost-vertex DSS: for each unique ghost vertex, sum the values of all its local vertices and store the sum at each of those local vertex locations in perimeter_data.

Part of ClimaCore.Spaces.weighted_dss!.

ClimaCore.Topologies.dss_ghost! — Function

dss_ghost!(device, perimeter_data, perimeter, topology)

Set the value in perimeter_data of all local vertices of each unique ghost vertex to that of the representative vertex.

Part of ClimaCore.Spaces.weighted_dss!.

ClimaCore.Topologies.create_dss_buffer — Function

create_dss_buffer(data, space)

Create a Topologies.DSSBuffer for the field data data on space. Return nothing if space is discontinuous (DG) or has no second horizontal dimension (1D spectral element spaces use dss_1d! and need no buffer).

Spaces.create_dss_buffer(fv::FieldVector)

Create a NamedTuple mapping the name of each field in fv to its DSS buffer.

Spaces.create_dss_buffer(field::Field)

Create a DSS buffer for communicating the neighbor information of field, or nothing if field needs no buffer.

ClimaCore.Topologies.fill_send_buffer! — Function

fill_send_buffer!(topology, data, ghost_buffer::GhostBuffer)

Load the send buffer of ghost_buffer with the data of the elements that neighboring processes need for their ghost elements. Returns nothing.

fill_send_buffer!(device, dss_buffer)

Load the send buffer of dss_buffer from its perimeter_data. For unique ghost vertices, only the representative vertices, which hold the result of the "ghost local" DSS, are loaded.

Part of ClimaCore.Spaces.weighted_dss!.

ClimaCore.Topologies.DSSBuffer — Type

DSSBuffer{T, G, D, A, B, VI}

Buffers and index maps for the direct stiffness summation (DSS) of one data layout with element type T on a Topology2D. Construct with create_dss_buffer(data, topology, local_geometry = nothing, dss_weights = nothing).

Fields

graph_context: ClimaComms graph context for communication.

perimeter_data: DataLayout holding the (transformed) values at the perimeter nodes of each element.

send_data: Send buffer, an AbstractVector{FT}.

recv_data: Receive buffer, an AbstractVector{FT}.

send_buf_idx: Indexing array for loading the send buffer from perimeter_data.

recv_buf_idx: Indexing array for adding the receive buffer into perimeter_data.

internal_elems: Local elements (lidx) that touch no ghost element.

perimeter_elems: Local elements (lidx) on the process boundary.

ClimaCore.Topologies.load_from_recv_buffer! — Function

load_from_recv_buffer!(device, dss_buffer)

Add the data in the receive buffer of dss_buffer to the corresponding locations in its perimeter_data. For ghost vertices, the data is added only to the representative vertices; dss_ghost! then scatters the values to the other local vertices of each unique ghost vertex.

Part of ClimaCore.Spaces.weighted_dss!.

ClimaCore.Topologies.dss! — Function

dss!(data, topology)

Perform unweighted DSS of data in place: values at nodes shared between elements are replaced by their sum. Returns nothing.

ClimaCore.Spaces.weighted_dss_start! — Function

weighted_dss_start!(data, space, dss_buffer)

Start the weighted DSS of data: prepare the perimeter data of the elements on the process boundary and begin communication with neighboring processes. Returns nothing.

It consists of the following steps:

Apply Spaces.dss_transform! on perimeter elements. This weights, and transforms to the physical basis if needed, the vector fields; scalar fields are weighted. The result is stored in dss_buffer.perimeter_data.

Apply Spaces.dss_local_ghost!, which computes the partial weighted DSS on ghost vertices using only the information from local vertices.

Apply Spaces.fill_send_buffer!, which loads the send buffer from perimeter_data. For unique ghost vertices, only the representative vertices, which hold the result of the "ghost local" DSS, are loaded.

Start the DSS communication with neighboring processes.

ClimaCore.Spaces.weighted_dss_internal! — Function

weighted_dss_internal!(data, space, dss_buffer)

Perform the part of the weighted DSS of data that needs no communication, while the communication started by Spaces.weighted_dss_start! is in flight. Returns nothing.

It consists of the following steps:

Apply Spaces.dss_transform! on interior elements. Local elements are split into interior and perimeter elements so that communication overlaps with computation.

Apply Spaces.dss_local!, which computes the weighted DSS on local vertices and faces.

Apply Spaces.dss_untransform! on interior elements.

On a 1D spectral element space, the whole DSS is performed here with dss_1d!.

ClimaCore.Spaces.weighted_dss_ghost! — Function

weighted_dss_ghost!(data, space, dss_buffer)

Finish the weighted DSS of data started by Spaces.weighted_dss_start!. Returns data.

It consists of the following steps:

Finish the communication.

Apply Spaces.load_from_recv_buffer!, which adds the data in the receive buffer to the corresponding locations in perimeter_data. For ghost vertices, the data is added only to the representative vertices; the values are then scattered to the other local vertices of each unique ghost vertex by dss_ghost!.

Apply Spaces.dss_untransform! on perimeter elements, which transforms the summed vectors back to their original basis and copies the summed data from perimeter_data to data.

ClimaCore.Spaces.weighted_dss! — Function

weighted_dss!(data, space, dss_buffer)

Compute the weighted direct stiffness summation (DSS) of data in place: values at nodes shared between elements are replaced by their weighted average. On a discontinuous (DG) space this is a no-op.

It consists of the following steps:

Spaces.weighted_dss_start!,

Spaces.weighted_dss_internal!,

Spaces.weighted_dss_ghost!.

Spaces.weighted_dss!(fv::FieldVector, dss_buffer = Spaces.create_dss_buffer(fv))

Apply weighted direct stiffness summation (DSS) to each field in fv in place. If dss_buffer is not provided, a buffer is created for each field in fv. The fields are passed to the Pair method of Spaces.weighted_dss!, which overlaps their communication.

Spaces.weighted_dss!(f::Field, dss_buffer = Spaces.create_dss_buffer(field))

Apply weighted direct stiffness summation (DSS) to f in place and return f. dss_buffer holds the buffers for communication in a distributed setting; see Spaces.create_dss_buffer. On a discontinuous (DG) space this is a no-op.

This is a projection operation from the piecewise polynomial space $V_0$ to the continuous space $V_1 = V_0 \cap C_0$, defined as the field $\theta \in V_1$ such that for all $\phi \in V_1$

\[\int_\Omega \phi \theta \,d\Omega = \int_\Omega \phi f \,d\Omega\]

In matrix form, we define $\bar \theta$ to be the unique global node representation, and $Q$ to be the "scatter" operator which maps to the redundant node representation $\theta$

\[\theta = Q \bar \theta\]

Then the problem can be written as

\[(Q \bar\phi)^\top W J Q \bar\theta = (Q \bar\phi)^\top W J f\]

which reduces to

\[\theta = Q \bar\theta = Q (Q^\top W J Q)^{-1} Q^\top W J f\]

Spaces.weighted_dss!(field1 => dss_buffer1, field2 => dss_buffer2, ...)

Call Spaces.weighted_dss! on multiple fields at once, overlapping the communication of all fields with computation. Returns nothing.

ClimaCore.Spaces.unique_nodes — Function

unique_nodes(space::SpectralElementSpace2D)

Return an iterator over the unique nodes of space. Each node is represented by the first ((i, j), e) triple at which it occurs.

This function is experimental, and may change in future.