DataLayouts

ClimaCore.DataLayouts — Module

ClimaCore.DataLayouts

Defines the following DataLayouts (see individual docs for more info):

TODO: Add links to these datalayouts

IJKFVH

IJFH

IJHF

IFH

IHF

DataF

IJF

IF

VF

VIJFH

VIJHF

VIFH

VIHF

IH1JH2

IV1JH2

Notation:

i,j are horizontal node indices within an element

k is the vertical node index within an element

f is the field index (1 if field is scalar, >1 if it is a vector field)

v is the vertical element index in a stack

h is the element stack index

Data layout is specified by the order in which they appear, e.g. IJKFVH indexes the underlying array as [i,j,k,f,v,h]

Datalayouts that end with the field index

One of the fundamental features of datalayouts is to be able to store multiple variables in the same array, and then access those variables by name. As such, we occasionally must index into multiple variables when performing operations with a datalayout.

We can efficiently support linear indexing with datalayouts whose field index (f) is first or last. This is for the same reason as https://docs.julialang.org/en/v1/devdocs/subarrays/#Linear-indexing:

Linear indexing can be implemented efficiently when the entire array
has a single stride that separates successive elements, starting from
some offset.

Therefore, we provide special handling for these datalayouts where possible to leverage efficient linear indexing.

Here are some references containing relevant discussions and efforts to leverage efficient linear indexing:

https://github.com/CliMA/ClimaCore.jl/issues/1889

https://github.com/JuliaLang/julia/issues/28126

https://github.com/JuliaLang/julia/issues/32051

https://github.com/maleadt/StaticCartesian.jl

https://github.com/JuliaGPU/GPUArrays.jl/pull/454#issuecomment-1431575721

https://github.com/JuliaGPU/GPUArrays.jl/pull/520

https://github.com/JuliaGPU/GPUArrays.jl/pull/464

ClimaCore.DataLayouts.DataF — Type

DataF{S, A} <: Data0D{S}

Backing DataLayout for 0D point data.

DataF{S}(ArrayType[, ones | zeros | rand])

The ArrayType constructor returns a DataF given the ArrayType and (optionally) an initialization method (one of Base.ones, Base.zeros, Random.rand).

ClimaCore.DataLayouts.IF — Type

IF{S, Ni, A} <: DataSlab1D{S, Ni}

Backing DataLayout for 1D spectral element slab data.

Nodal element data (I) are contiguous for each S datatype struct field (F) for a single element slab.

A DataSlab1D view can be returned from other Data1D objects by calling slab(data, idx...).

IF{S}(ArrayType[, ones | zeros | rand]; Ni)

The keyword constructor returns a IF given the ArrayType and (optionally) an initialization method (one of Base.ones, Base.zeros, Random.rand) and the keywords:

Ni quadrature degrees of freedom in the horizontal direction

Note

Objects made with the keyword constructor accept integer keyword inputs, so they are dynamically created. You may want to use a different constructor if you're making the object in a performance-critical section, and if you know the type parameters at compile time.

ClimaCore.DataLayouts.IJF — Type

IJF{S, Nij, A} <: DataSlab2D{S, Nij}

Backing DataLayout for 2D spectral element slab data.

Nodal element data (I,J) are contiguous for each S datatype struct field (F) for a single element slab.

A DataSlab2D view can be returned from other Data2D objects by calling slab(data, idx...).

IJF{S}(ArrayType[, ones | zeros | rand]; Nij)

The keyword constructor returns a IJF given the ArrayType and (optionally) an initialization method (one of Base.ones, Base.zeros, Random.rand) and the keywords:

Nij quadrature degrees of freedom per horizontal direction

Note

Objects made with the keyword constructor accept integer keyword inputs, so they are dynamically created. You may want to use a different constructor if you're making the object in a performance-critical section, and if you know the type parameters at compile time.

ClimaCore.DataLayouts.VF — Type

VF{S, A} <: DataColumn{S, Nv}

Backing DataLayout for 1D FV column data.

Column level data (V) are contiguous for each S datatype struct field (F).

A DataColumn view can be returned from other Data1DX, Data2DX objects by calling column(data, idx...).

VF{S}(ArrayType[, ones | zeros | rand]; Nv)

The keyword constructor returns a VF given the ArrayType and (optionally) an initialization method (one of Base.ones, Base.zeros, Random.rand) and the keywords:

Nv number of vertical degrees of freedom

Note

Objects made with the keyword constructor accept integer keyword inputs, so they are dynamically created. You may want to use a different constructor if you're making the object in a performance-critical section, and if you know the type parameters at compile time.

ClimaCore.DataLayouts.IFH — Type

IFH{S,Ni,Nh,A} <: Data1D{S, Ni}
IFH{S,Ni,Nh}(ArrayType)

Backing DataLayout for 1D spectral element slabs.

Element nodal point (I) data is contiguous for each datatype S struct field (F), for each 1D mesh element (H).

The ArrayType-constructor makes a IFH 1D Spectral DataLayout given the backing ArrayType, quadrature degrees of freedom Ni, and the number of mesh elements Nh.

IFH{S}(ArrayType[, ones | zeros | rand]; Ni, Nh)

The keyword constructor returns a IFH given the ArrayType and (optionally) an initialization method (one of Base.ones, Base.zeros, Random.rand) and the keywords:

Ni quadrature degrees of freedom in the horizontal direction

Nh number of mesh elements

Note

Objects made with the keyword constructor accept integer keyword inputs, so they are dynamically created. You may want to use a different constructor if you're making the object in a performance-critical section, and if you know the type parameters at compile time.

ClimaCore.DataLayouts.IJFH — Type

IJFH{S, Nij, A} <: Data2D{S, Nij}
IJFH{S,Nij}(ArrayType, nelements)

Backing DataLayout for 2D spectral element slabs.

Element nodal point (I,J) data is contiguous for each datatype S struct field (F), for each 2D mesh element slab (H).

The ArrayType-constructor constructs a IJFH 2D Spectral DataLayout given the backing ArrayType, quadrature degrees of freedom Nij × Nij, and the number of mesh elements nelements.

IJFH{S}(ArrayType[, Base.ones | zeros | rand]; Nij, Nh)

The keyword constructor returns a IJFH given the ArrayType and (optionally) an initialization method (one of Base.ones, Base.zeros, Random.rand) and the keywords:

Nij quadrature degrees of freedom per horizontal direction

Nh number of mesh elements

Note

Objects made with the keyword constructor accept integer keyword inputs, so they are dynamically created. You may want to use a different constructor if you're making the object in a performance-critical section, and if you know the type parameters at compile time.

ClimaCore.DataLayouts.VIFH — Type

VIFH{S, Nv, Ni, A} <: Data1DX{S, Nv, Ni}

Backing DataLayout for 1D spectral element slab + extruded 1D FV column data.

Column levels (V) are contiguous for every element nodal point (I) for each datatype S struct field (F), for each 1D mesh element slab (H).

VIFH{S}(ArrayType[, ones | zeros | rand]; Nv, Ni, Nh)

The keyword constructor returns a VIFH given the ArrayType and (optionally) an initialization method (one of Base.ones, Base.zeros, Random.rand) and the keywords:

Nv number of vertical degrees of freedom

Ni quadrature degrees of freedom in the horizontal direction

Nh number of horizontal elements

Note

Objects made with the keyword constructor accept integer keyword inputs, so they are dynamically created. You may want to use a different constructor if you're making the object in a performance-critical section, and if you know the type parameters at compile time.

ClimaCore.DataLayouts.VIJFH — Type

VIJFH{S, Nij, A} <: Data2DX{S, Nij}

Backing DataLayout for 2D spectral element slab + extruded 1D FV column data.

Column levels (V) are contiguous for every element nodal point (I, J) for each S datatype struct field (F), for each 2D mesh element slab (H).

VIJFH{S}(ArrayType[, ones | zeros | rand]; Nv, Nij, Nh)

The keyword constructor returns a VIJFH given the ArrayType and (optionally) an initialization method (one of Base.ones, Base.zeros, Random.rand) and the keywords:

Nv number of vertical degrees of freedom

Nij quadrature degrees of freedom per horizontal direction

Nh number of horizontal elements

Note

Objects made with the keyword constructor accept integer keyword inputs, so they are dynamically created. You may want to use a different constructor if you're making the object in a performance-critical section, and if you know the type parameters at compile time.

ClimaCore.DataLayouts.IHF — Type

IHF{S,Ni,Nh,A} <: Data1D{S, Ni}
IHF{S,Ni,Nh}(ArrayType)

Backing DataLayout for 1D spectral element slabs.

Element nodal point (I) data is contiguous for each datatype S struct field (F), for each 1D mesh element (H).

The ArrayType-constructor makes a IHF 1D Spectral DataLayout given the backing ArrayType, quadrature degrees of freedom Ni, and the number of mesh elements Nh.

IHF{S}(ArrayType[, ones | zeros | rand]; Ni, Nh)

The keyword constructor returns a IHF given the ArrayType and (optionally) an initialization method (one of Base.ones, Base.zeros, Random.rand) and the keywords:

Ni quadrature degrees of freedom in the horizontal direction

Nh number of mesh elements

Note

Objects made with the keyword constructor accept integer keyword inputs, so they are dynamically created. You may want to use a different constructor if you're making the object in a performance-critical section, and if you know the type parameters at compile time.

ClimaCore.DataLayouts.IJHF — Type

IJHF{S, Nij, A} <: Data2D{S, Nij}
IJHF{S,Nij}(ArrayType, nelements)

Backing DataLayout for 2D spectral element slabs.

Element nodal point (I,J) data is contiguous for each datatype S struct field (F), for each 2D mesh element slab (H).

The ArrayType-constructor constructs a IJHF 2D Spectral DataLayout given the backing ArrayType, quadrature degrees of freedom Nij × Nij, and the number of mesh elements nelements.

IJHF{S}(ArrayType[, Base.ones | zeros | rand]; Nij, Nh)

The keyword constructor returns a IJHF given the ArrayType and (optionally) an initialization method (one of Base.ones, Base.zeros, Random.rand) and the keywords:

Nij quadrature degrees of freedom per horizontal direction

Nh number of mesh elements

Note

Objects made with the keyword constructor accept integer keyword inputs, so they are dynamically created. You may want to use a different constructor if you're making the object in a performance-critical section, and if you know the type parameters at compile time.

ClimaCore.DataLayouts.VIHF — Type

VIHF{S, Nv, Ni, A} <: Data1DX{S, Nv, Ni}

Backing DataLayout for 1D spectral element slab + extruded 1D FV column data.

Column levels (V) are contiguous for every element nodal point (I) for each datatype S struct field (F), for each 1D mesh element slab (H).

VIHF{S}(ArrayType[, ones | zeros | rand]; Nv, Ni, Nh)

The keyword constructor returns a VIHF given the ArrayType and (optionally) an initialization method (one of Base.ones, Base.zeros, Random.rand) and the keywords:

Nv number of vertical degrees of freedom

Ni quadrature degrees of freedom in the horizontal direction

Nh number of horizontal elements

Note

Objects made with the keyword constructor accept integer keyword inputs, so they are dynamically created. You may want to use a different constructor if you're making the object in a performance-critical section, and if you know the type parameters at compile time.

ClimaCore.DataLayouts.VIJHF — Type

VIJHF{S, Nij, A} <: Data2DX{S, Nij}

Backing DataLayout for 2D spectral element slab + extruded 1D FV column data.

Column levels (V) are contiguous for every element nodal point (I, J) for each S datatype struct field (F), for each 2D mesh element slab (H).

VIJHF{S}(ArrayType[, ones | zeros | rand]; Nv, Nij, Nh)

The keyword constructor returns a VIJHF given the ArrayType and (optionally) an initialization method (one of Base.ones, Base.zeros, Random.rand) and the keywords:

Nv number of vertical degrees of freedom

Nij quadrature degrees of freedom per horizontal direction

Nh number of horizontal elements

Note

Objects made with the keyword constructor accept integer keyword inputs, so they are dynamically created. You may want to use a different constructor if you're making the object in a performance-critical section, and if you know the type parameters at compile time.

ClimaCore.DataLayouts.parent_array_type — Function

parent_array_type(A, [B])

Determines the array type underlying the wrapper type A, dropping all parameters related to array dimensions. A new basetype B can be specified to replace the original eltype(A).

parent_array_type(data::AbstractData)

This is an internal function, please do not use outside of ClimaCore.

Returns the the backing array type.

This function is helpful for writing generic code, when reconstructing new datalayouts with new type parameters.

ClimaCore.DataLayouts.promote_parent_array_type — Function

promote_parent_array_type(A1, A2)

Promotes two array types A1 and A2 generated by parent_array_type, which includes promoting the basetypes eltype(A1) and eltype(A2).

ClimaCore.DataLayouts.default_basetype — Function

default_basetype(T)

Recursively searches through the fieldtypes of T for a type that set_struct! and get_struct can use to represent all other types nested in T. If no such type is found, an unsigned integer type is selected based on the size of T.

ClimaCore.DataLayouts.check_basetype — Function

check_basetype(B, T)

Checks if set_struct! and get_struct can use values of type B to store values of type T. Throws an error if B is not a valid basetype, printing out one fieldtype of T that cannot use B as a basetype.

ClimaCore.DataLayouts.checked_valid_basetype — Function

checked_valid_basetype(B, T)

Returns either B or the default_basetype of T, depending on whether B is a valid basetype for T.

ClimaCore.DataLayouts.storage_length — Function

storage_length(B, T)

Determines how many values of type B are required by set_struct! and get_struct to store a single value of type T.

ClimaCore.DataLayouts.struct_field_view — Function

struct_field_view(array, T, Val(F), [Val(D)])

Creates a view of the data in array that corresponds to a particular field of T, assuming that array has been populated by set_struct!. The field is specified through a Val that contains its index F, and it can be loaded from the resulting view using get_struct.

For multidimensional arrays with values of type T stored along a particular dimension, the resulting view contains the specified field from each value. By default, it is assumed that each value is stored along the last dimension, but any other dimension can be specified through a Val containing its index D.

ClimaCore.DataLayouts.set_struct! — Function

set_struct!(array, struct_val, [start], [Val(D)])

Populates array with data that represents struct_val, using bitcast_struct to convert struct_val into elements of the array.

For multidimensional arrays with values of type T stored along a particular dimension, a start index is used to specify where to store struct_val. By default, each value is stored along the last dimension, but any other dimension can be specified through a Val containing its index D.

Examples

julia> set_struct!(zeros(Int8, 4), Int32(1))
4-element Vector{Int8}:
 1
 0
 0
 0

julia> set_struct!(zeros(Int64, 3), (Int32(2), Int32(0), Int128(1)))
3-element Vector{Int64}:
 2
 1
 0

julia> set_struct!(zeros(Int64, 2, 3), (Int32(2), Int32(0), Int128(1)), 2)
2×3 Matrix{Int64}:
 0  0  0
 2  1  0

ClimaCore.DataLayouts.get_struct — Function

get_struct(array, T, [start], [Val(D)])

Loads a value of type T that set_struct! has stored in array, using bitcast_struct to convert elements of the array into a value of type T.

As with set_struct!, a start index and an optional dimension index D can be used to select a single value from an array that contains multiple values.

Examples

julia> get_struct(Int8[1, 0, 0, 0], Int32)
1

julia> get_struct([2, 1, 0], Tuple{Int32, Int32, Int128})
(2, 0, 1)

julia> get_struct([0 0 0; 2 1 0], Tuple{Int32, Int32, Int128}, 2)
(2, 0, 1)

ClimaCore.DataLayouts.bitcast_struct — Function

bitcast_struct(T, struct_val)

Converts a data structure to any isbits type T that spans the same number of bytes, generalizing Julia's native Core.bitcast function to composite types.

Examples

julia> bitcast_struct(NTuple{4, Int8}, Int32(1))
(1, 0, 0, 0)

julia> bitcast_struct(NTuple{6, Int32}, (2 * eps(0.0), eps(0.0), 0.0))
(2, 0, 1, 0, 0, 0)

julia> bitcast_struct(Tuple{Int32, Int32, Int128}, (2, 1, 0))
(2, 0, 1)

Extended help

The output of bitcast_struct is identical to reinterpret(T, struct_val), with both functions interpreting sequential bytes of data in little-endian order:

julia> reinterpret(NTuple{4, Int8}, Int32(1))
(1, 0, 0, 0)

julia> reinterpret(NTuple{6, Int32}, (2 * eps(0.0), eps(0.0), 0.0))
(2, 0, 1, 0, 0, 0)

julia> reinterpret(Tuple{Int32, Int32, Int128}, (2, 1, 0))
(2, 0, 1)

Although their outputs are identical, bitcast_struct and reinterpret have very different implementations. Unlike bitcast_struct, which performs all conversions in register memory by calling Core.bitcast, reinterpret only does this for isprimitivetype inputs, and otherwise it converts by allocating Refs and passing them to unsafe_convert. Since the heap memory where Refs are allocated does not exist on GPUs, reinterpret cannot be used to convert composite types inside GPU kernels, and bitcast_struct must be used instead.

It is also important to note that bitcast_struct is only similar to the method of reinterpret for isbits inputs, rather than the more common method for AbstractArray inputs. These methods differ in how they handle padding, which the C code underlying Julia uses to ensure that fields of different sizes within the same data structure are aligned in register memory. While the array form of reinterpret only converts between types with the same amount of padding, the isbits form is able to support arbitrary padding, so it can store a padded value of type T in an array smaller than sizeof(T) (the smallest possible size of a nonempty Array{T} or ReinterpretArray{T}).

The previous example of converting Int64s into a Tuple{Int32, Int32, Int128} illustrates this difference. The isbits form of reinterpret only needs three Int64s to perform the conversion, but the array form needs a fourth Int64, spanning the eight bytes of padding used to align the Int32s and the Int128:

julia> reinterpret(Tuple{Int32, Int32, Int128}, (2, 1, 0))
(2, 0, 1)

julia> reinterpret(reshape, Tuple{Int32, Int32, Int128}, [2, 1, 0])[1]
ERROR: ArgumentError: [...]

julia> reinterpret(reshape, Tuple{Int32, Int32, Int128}, [2, 0, 1, 0])[1]
(2, 0, 1)

For more information about reinterpret and padding, see the following:

https://discourse.julialang.org/t/reinterpret-returns-wrong-values

https://discourse.julialang.org/t/reinterpret-vector-into-single-struct

https://discourse.julialang.org/t/reinterpret-vector-of-mixed-type-tuples