Utilities

Type Functions

ClimaCore.Utilities.unionall_type — Function

unionall_type(::Type{T})

Extract the type of the input, and strip it of any type parameters.

This is useful when one needs the generic constructor for a given type.

Example

julia> unionall_type(typeof([1, 2, 3]))
Array

julia> struct Foo{A, B}
               a::A
               b::B
       end

julia> unionall_type(typeof(Foo(1,2)))
Foo

ClimaCore.Utilities.replace_type_parameter — Function

replace_type_parameter(T, P, P′)

Recursively modifies the parameters of T, replacing every subtype of P with P′. This is like constructing a value of type T and converting subfields of type P to type P′, though no constructors are actually called or compiled.

ClimaCore.Utilities.fieldtype_vals — Function

fieldtype_vals(T)

Statically inferrable analogue of Val.(fieldtypes(T)). Functions of Types are specialized upon successful constant propagation, but functions of Vals are always specialized, so fieldtype_vals can be used in place of fieldtypes to ensure that recursive functions over nested types have inferrable outputs.

ClimaCore.Utilities.new — Function

new(T, [fields])

Exposes the new pseudo-function that allocates a value of type T with specific fields. Unlike the pseudo-function, this does not automatically convert the given fields to the fieldtypes of T. The second argument can also be omitted to leave the fields of the allocated value uninitialized.

ClimaCore.Utilities.unsafe_eltype — Function

unsafe_eltype(itr)

Analogue of eltype with support for un-materialized broadcast expressions, adapted from Base.Broadcast.combine_eltypes. Does not perform any safety checks, and may potentially return non-concrete types (like an empty Union{}).

ClimaCore.Utilities.safe_eltype — Function

safe_eltype(itr)

Analogue of eltype with support for un-materialized broadcast expressions, adapted from Base.Broadcast.combine_eltypes. Throws an error when the concrete element type of a broadcast expression cannot be inferred, indicating which part of the expression triggers the first type instability during inference.

Utilities.AutoBroadcaster

ClimaCore.Utilities.AutoBroadcaster — Type

AutoBroadcaster(itr)

Wrapper for an iterator that forces certain functions to be broadcasted over the iterator's elements. This allows different types of broadcasting to be applied simultaneously; e.g., ClimaCore's Fields and similar types use the standard dot syntax to denote parallelized iteration over spatial locations, and they wrap their values in AutoBroadcasters for unrolled iteration over subfields. All statically-sized iterators for which is_auto_broadcastable is true are compatible with AutoBroadcasters.

In the context of AutoBroadcasters, broadcasting a function applies it with unrolled_map, iterating over all arguments for which is_auto_broadcastable is true (including those not wrapped in AutoBroadcasters), while other arguments are passed to the function directly. This behavior is triggered by using AutoBroadcasters, optionally in conjunction with compatible iterators that are not wrapped in AutoBroadcasters, in the following ways:

passing them to standard math functions or constructors

passing them to ifelse (for iterating over conditional values)

applying them as function calls (for iterating over functions)

explicitly broadcasting over them with nested_auto_broadcast or @.

Nested AutoBroadcasters constructed with add_auto_broadcasters evaluate broadcasts recursively, mapping across every layer of nested iterators so that broadcasted functions are only applied to non-iterators in the innermost layers. Aside from automatic broadcasting, AutoBroadcasters are essentially identical to their underlying iterators, with support for common operations like iterate, propertynames, getindex, and reduce.

Examples

julia> x = Utilities.AutoBroadcaster((1, 2.0, StaticArrays.SVector(3, 4)))
(1, 2.0, [3, 4])

julia> zero(typeof(x))
(0, 0.0, [0, 0])

julia> 2 * x - (2, 3, [4, 5])
(0, 1.0, [2, 3])

julia> y = Utilities.add_auto_broadcasters((1, 2, (a = 3, b = 4, c = (5, 6, (7, 8)))))
(1, 2, (a = 3, b = 4, c = (5, 6, (7, 8))))

julia> min(y, abs(5 - y))
(1, 2, (a = 2, b = 1, c = (0, 1, (2, 3))))

julia> x' * y * x ÷ 5
(0, 1.0, (a = 15, b = 20, c = (25, 30, (35, 40))))

ClimaCore.Utilities.is_auto_broadcastable — Function

is_auto_broadcastable(::Type)
is_auto_broadcastable(itr)

Indicates whether an AutoBroadcaster should broadcast over iterators of the given type. By default, this is only true for Tuple and NamedTuple types, but it can be extended to any statically-sized type compatible with UnrolledUtilities.jl.

For convenience, is_auto_broadcastable also supports passing a concrete iterator instead of its type, but this method should not be extended directly.

ClimaCore.Utilities.add_auto_broadcasters — Function

add_auto_broadcasters(itr)
add_auto_broadcasters(I)

Recursively applies the AutoBroadcaster constructor to iterators for which is_auto_broadcastable is true, as well as their elements for which it is true, while leaving values for which it is false unmodified. Can also be passed an iterator's type to infer the result type for such an iterator.

ClimaCore.Utilities.drop_auto_broadcasters — Function

drop_auto_broadcasters(x)
drop_auto_broadcasters(X)

Recursively unwraps constructors applied by add_auto_broadcasters, extracting the iterator from every AutoBroadcaster in x. Can also be passed an iterator's type to infer the result type for such an iterator.

ClimaCore.Utilities.EnableAutoBroadcasting — Type

EnableAutoBroadcasting(f)

Wrapper for a function f that calls either add_auto_broadcasters or drop_auto_broadcasters on its arguments if doing so may prevent an error from being thrown.

When EnableAutoBroadcasting(f)(args...) is evaluated, if the return_type of f(args...) indicates that an error will be thrown, the return_type is recomputed after add_auto_broadcasters is called on every argument, and then again after drop_auto_broadcasters is called on every argument. If either of the new results no longer corresponds to a guaranteed error, f is applied to the modified arguments. Otherwise, f is applied to its original arguments.

ClimaCore.Utilities.nested_auto_broadcast — Function

nested_auto_broadcast(f, args...)

Analogue of broadcast that is applied recursively over nested iterators, as long as at least one argument is an AutoBroadcaster. This is always the case for broadcast expressions that only contain AutoBroadcasters and scalars, arrays, or tuples; otherwise, this function must be used instead.

Utilities.PlusHalf

ClimaCore.Utilities.PlusHalf — Type

PlusHalf(i)

Represents i + 1/2, but stored as internally as an integer value. Used for indexing into staggered finite difference meshes: the convention "half" values are indexed at cell faces, whereas centers are indexed at cell centers.

Supports +, - and inequalities.

See also half.

ClimaCore.Utilities.half — Constant

const half = PlusHalf(0)

Utilities.Cache

ClimaCore.Utilities.Cache — Module

Utilities.Cache

ClimaCore maintains an internal cache of topology and grid objects: this ensures that if the constructor with the same arguments is invoked again (e.g. by reading from a file), the cached object will be returned (also known as memoization). This has two main advantages:

topology and metric information can be reused, reducing memory usage.

it is easy to check if two fields live on the same grid: we can just check if the underlying grid objects are the same (===), rather than checking all the fields are equal (via ==).

However this means that objects in the cache will not be removed from the garbage collector, so we provide an interface to remove these.

ClimaCore.Utilities.Cache.cached_objects — Function

Utilities.Cache.cached_objects()

List all currently cached objects.

ClimaCore.Utilities.Cache.clean_cache! — Function

Utilities.Cache.clean_cache!(object)

Remove object from the cache of created objects.

In most cases, this function should not need to be called, unless you are constructing many grid objects, for example when doing a sweep over grid paramaters.

Utilities.Cache.clean_cache!()

Remove all objects from the cache of created objects.

In most cases, this function should not need to be called, unless you are constructing many grid objects, for example when doing a sweep over grid paramaters.