Utilities

ClimaCore.Utilities.PlusHalf — Type

PlusHalf(i)

Represents i + 1/2, but stored as internally as an integer value. Used for indexing into staggered finite difference meshes: the convention "half" values are indexed at cell faces, whereas centers are indexed at cell centers.

Supports +, - and inequalities.

See also half.

ClimaCore.Utilities.half — Constant

const half = PlusHalf(0)

Utilities.MathMapper

ClimaCore.Utilities.MathMapper — Type

MathMapper(itr)

Wrapper for an iterator that forces math functions (+, *, zero, etc.) to be broadcasted over the iterator's elements. Aside from automatic broadcasting, MathMappers act exactly like their underlying iterators, with support for common operations like propertynames, getindex, getindex, and reduce. All statically-sized iterators for which supports_math_mapper is true are compatible with MathMappers.

In the context of MathMappers, broadcasting a function corresponds to mapping it over all arguments for which supports_math_mapper is true (including those not wrapped in MathMappers), while other arguments are passed to it directly. This behavior is triggered when a MathMapper is used in the following ways:

passed to map or mapreduce (only up to 4 arguments)

passed to a math function from Base (all common single-output functions)

passed to ifelse (for mapping over conditional expressions)

applied as a function call (for mapping over multiple functions)

Nested MathMappers evaluate broadcasts recursively, mapping down through every nested layer and only applying broadcasted functions to non-MathMappers in the innermost layers. Reductions of nested MathMappers are similarly evaluated over non-MathMappers in the innermost layers. The MathMapper constructor only wraps outer layers of nested iterators, but nested_math_mapper recursively wraps all layers that support MathMappers.

Examples

julia> x = Utilities.MathMapper((1, 2.0, StaticArrays.SVector(3, 4)))
(1, 2.0, [3, 4])

julia> zero(typeof(x))
(0, 0.0, [0, 0])

julia> 2 * x - (2, 3, [4, 5])
(0, 1.0, [2, 3])

julia> y = Utilities.nested_math_mapper((a = 1, b = (2, 3, (d = 4, e = (5,)))))
(a = 1, b = (2, 3, (d = 4, e = (5,))))

julia> min(y, abs(4 - y))
(a = 1, b = (2, 1, (d = 0, e = (1,))))

julia> mapreduce(min, +, y, abs(4 - y))
5

julia> maximum(log2, y - one(y))
2.0

ClimaCore.Utilities.supports_math_mapper — Function

supports_math_mapper(itr)

Indicates whether a MathMapper should map over the given iterator. By default, this is true for all Tuples and NamedTuples.

Defining the method supports_math_mapper(::T) = true allows MathMappers to map over iterators of any type T, as long as those iterators are statically sized and compatible with the unrolled functions in UnrolledUtilities.jl.

ClimaCore.Utilities.nested_math_mapper — Function

nested_math_mapper(itr)

Recursively applies the MathMapper constructor to an iterator and all of its elements for which supports_math_mapper is true. Values for which supports_math_mapper is false are left unmodified.

ClimaCore.Utilities.nested_math_mapper_type — Function

nested_math_mapper_type(T)

Type of a nested_math_mapper constructed from an iterator of type T.

ClimaCore.Utilities.math_mapper_broadcast — Function

math_mapper_broadcast(f, args...)

Generic form of map(f, args...) for MathMapper arguments, which performs automatic broadcasting with recursion over nested iterators. The built-in map only uses methods for MathMappers when given arguments of specific types, but this function works with any arguments.

ClimaCore.Utilities.math_mapper_type_broadcast — Function

math_mapper_type_broadcast(f, types...)

Similar to math_mapper_broadcast, but for argument types rather than the arguments themselves. Instead of applying f to elements of arguments for which supports_math_mapper is true, this function applies f to the types of those elements (i.e., to the fieldtypes of the argument types).

ClimaCore.Utilities.reduce_math_mapper_broadcast — Function

reduce_math_mapper_broadcast(f, op, args...; init...)

Generic form of mapreduce(f, op, args...; init...) for MathMapper arguments, which performs automatic broadcasting with recursion over nested iterators. The built-in mapreduce only uses methods for MathMappers when given arguments of specific types, but this function works with any arguments.

ClimaCore.Utilities.reduce_math_mapper_type_broadcast — Function

reduce_math_mapper_type_broadcast(f, types...)

Similar to reduce_math_mapper_broadcast, but for argument types rather than the arguments themselves. Instead of applying f to elements of arguments for which supports_math_mapper is true, this function applies f to the types of those elements (i.e., to the fieldtypes of the argument types).

ClimaCore.Utilities.@math_mapper_method — Macro

@math_mapper_method [T] <method definition>

Adds type constraints to all positional arguments in a method definition (not Vararg or keyword arguments) that do not already have constrained types, so that at least one argument is a MathMapper and the rest have type T.

If supports_math_mapper is true for values of type T, this ensures they will be promoted to behave like MathMappers in broadcast operations. If it is false, this instead allows them to be used as single values. By default, T is set to Union{Number, AbstractArray, Tuple, NamedTuple}, so that Tuple and NamedTuple iterators are mapped over while Number and AbstractArray values are treated like scalars.

To avoid ambiguity errors, a separate method must be defined for every distinct permutation of MathMappers and Ts, so that the number of methods actually generated by @math_mapper_method for $N$ type constraints is $2^N - 1$:

julia> @macroexpand @math_mapper_method NonMathMapper foo(x, y) = map(foo, x, y)
quote
    foo(x::MathMapper, y::MathMapper) = map(foo, x, y)
    foo(x::NonMathMapper, y::MathMapper) = map(foo, x, y)
    foo(x::MathMapper, y::NonMathMapper) = map(foo, x, y)
end

Utilities.Cache

ClimaCore.Utilities.Cache — Module

Utilities.Cache

ClimaCore maintains an internal cache of topology and grid objects: this ensures that if the constructor with the same arguments is invoked again (e.g. by reading from a file), the cached object will be returned (also known as memoization). This has two main advantages:

topology and metric information can be reused, reducing memory usage.

it is easy to check if two fields live on the same grid: we can just check if the underlying grid objects are the same (===), rather than checking all the fields are equal (via ==).

However this means that objects in the cache will not be removed from the garbage collector, so we provide an interface to remove these.

ClimaCore.Utilities.Cache.cached_objects — Function

Utilities.Cache.cached_objects()

List all currently cached objects.

ClimaCore.Utilities.Cache.clean_cache! — Function

Utilities.Cache.clean_cache!(object)

Remove object from the cache of created objects.

In most cases, this function should not need to be called, unless you are constructing many grid objects, for example when doing a sweep over grid paramaters.

Utilities.Cache.clean_cache!()

Remove all objects from the cache of created objects.

In most cases, this function should not need to be called, unless you are constructing many grid objects, for example when doing a sweep over grid paramaters.