Utilities

Type Functions

ClimaCore.Utilities.unionall_type — Function

unionall_type(::Type{T})

Extract the type of the input, and strip it of any type parameters.

This is useful when one needs the generic constructor for a given type.

Example

julia> unionall_type(typeof([1, 2, 3]))
Array

julia> struct Foo{A, B}
               a::A
               b::B
       end

julia> unionall_type(typeof(Foo(1,2)))
Foo

ClimaCore.Utilities.inferred_type — Function

inferred_type(f, types...)
inferred_type(f)

The type of value returned by f for arguments of the given types. An error is thrown if the concrete result type cannot be inferred, or if there is no method of f that can handle the specified types.

Omitting the type arguments causes inferred_type to generate the function closure (types...) -> inferred_type(f, types...).

ClimaCore.Utilities.inferred_const — Function

inferred_const(f, types...)
inferred_const(f)

The constant value returned by f for arguments of the given types, computed similarly to inferred_type, but with a Val wrapper. Only quantities that are marked as Core.Const by @code_warntype can be inferred like this.

Omitting the type arguments causes inferred_const to generate the function closure (types...) -> inferred_const(f, types...).

Utilities.AutoBroadcaster

ClimaCore.Utilities.AutoBroadcaster — Type

AutoBroadcaster(itr)

Wrapper for an iterator that forces certain functions to be broadcasted over the iterator's elements. This allows different types of broadcasting to be applied simultaneously; e.g., ClimaCore's Fields and similar types use the standard dot syntax to denote parallelized iteration over spatial locations, and they wrap their values in AutoBroadcasters for unrolled iteration over subfields. All statically-sized iterators for which is_auto_broadcastable is true are compatible with AutoBroadcasters.

In the context of AutoBroadcasters, broadcasting a function applies it with unrolled_map, iterating over all arguments for which is_auto_broadcastable is true (including those not wrapped in AutoBroadcasters), while other arguments are passed to the function directly. This behavior is triggered by using AutoBroadcasters, optionally in conjunction with compatible iterators that are not wrapped in AutoBroadcasters, in the following ways:

passing them to standard math functions or constructors

passing them to ifelse (for iterating over conditional values)

applying them as function calls (for iterating over functions)

manually broadcasting over them (only works with unwrapped iterators when broadcasting functions that have no more than 4 arguments)

Nested AutoBroadcasters constructed with enable_auto_broadcasting evaluate broadcasts recursively, mapping across every layer of nested iterators so that broadcasted functions are only applied to non-iterators in the innermost layers. Aside from automatic broadcasting, AutoBroadcasters are essentially identical to their underlying iterators, with support for common operations like iterate, propertynames, getindex, and reduce.

Examples

julia> x = Utilities.AutoBroadcaster((1, 2.0, StaticArrays.SVector(3, 4)))
(1, 2.0, [3, 4])

julia> zero(typeof(x))
(0, 0.0, [0, 0])

julia> 2 * x - (2, 3, [4, 5])
(0, 1.0, [2, 3])

julia> y = Utilities.enable_auto_broadcasting((1, 2, (a = 3, b = 4, c = (5, 6, (7, 8)))))
(1, 2, (a = 3, b = 4, c = (5, 6, (7, 8))))

julia> min(y, abs(5 - y))
(1, 2, (a = 2, b = 1, c = (0, 1, (2, 3))))

julia> x' * y * x ÷ 5
(0, 1.0, (a = 15, b = 20, c = (25, 30, (35, 40))))

ClimaCore.Utilities.is_auto_broadcastable — Function

is_auto_broadcastable(::Type)
is_auto_broadcastable(itr)

Indicates whether an AutoBroadcaster should broadcast over iterators of the given type. By default, this is only true for Tuple and NamedTuple types, but it can be extended to any statically-sized type compatible with UnrolledUtilities.jl.

For convenience, is_auto_broadcastable also supports passing a concrete iterator instead of its type, but this method should not be extended directly.

ClimaCore.Utilities.enable_auto_broadcasting — Function

enable_auto_broadcasting(itr)

Recursively applies the AutoBroadcaster constructor to iterators for which is_auto_broadcastable is true, as well as their elements for which it is true, while leaving values for which it is false unmodified.

ClimaCore.Utilities.disable_auto_broadcasting — Function

disable_auto_broadcasting(x)

Recursively unwraps constructors applied by enable_auto_broadcasting, extracting the iterator from every AutoBroadcaster in x.

ClimaCore.Utilities.auto_broadcast — Function

auto_broadcast(f)

Modifies a function so that it is preceded by enable_auto_broadcasting and followed by disable_auto_broadcasting every time it is called.

ClimaCore.Utilities.@auto_broadcaster_args — Macro

@auto_broadcaster_args [T] <method definition>

Adds type constraints to all positional arguments in a method definition (not Vararg or keyword arguments) that do not already have constrained types, so at least one argument is a AutoBroadcaster and the rest have type T.

If is_auto_broadcastable is true for a positional argument, that argument will behave like an AutoBroadcaster when broadcasting. Otherwise, that argument will be passed along as a single value. By default, T is a union over Tuples and NamedTuples, whose elements are mapped over, as well as Numbers and AbstractArrays, which are treated like single values.

To avoid ambiguity errors, a separate method must be defined for every distinct permutation of AutoBroadcasters and Ts. When adding $N$ type constraints, this macro generates $2^N - 1$ methods; e.g., for 2 new type constraints:

julia> @macroexpand @auto_broadcaster_args foo(x, y) = foo.(x, y)
quote
    foo(x::AutoBroadcaster, y::AutoBroadcaster) = foo.(x, y)
    foo(x::DefaultNonAutoBroadcaster, y::AutoBroadcaster) = foo.(x, y)
    foo(x::AutoBroadcaster, y::DefaultNonAutoBroadcaster) = foo.(x, y)
end

Utilities.PlusHalf

ClimaCore.Utilities.PlusHalf — Type

PlusHalf(i)

Represents i + 1/2, but stored as internally as an integer value. Used for indexing into staggered finite difference meshes: the convention "half" values are indexed at cell faces, whereas centers are indexed at cell centers.

Supports +, - and inequalities.

See also half.

ClimaCore.Utilities.half — Constant

const half = PlusHalf(0)

Utilities.Cache

ClimaCore.Utilities.Cache — Module

Utilities.Cache

ClimaCore maintains an internal cache of topology and grid objects: this ensures that if the constructor with the same arguments is invoked again (e.g. by reading from a file), the cached object will be returned (also known as memoization). This has two main advantages:

topology and metric information can be reused, reducing memory usage.

it is easy to check if two fields live on the same grid: we can just check if the underlying grid objects are the same (===), rather than checking all the fields are equal (via ==).

However this means that objects in the cache will not be removed from the garbage collector, so we provide an interface to remove these.

ClimaCore.Utilities.Cache.cached_objects — Function

Utilities.Cache.cached_objects()

List all currently cached objects.

ClimaCore.Utilities.Cache.clean_cache! — Function

Utilities.Cache.clean_cache!(object)

Remove object from the cache of created objects.

In most cases, this function should not need to be called, unless you are constructing many grid objects, for example when doing a sweep over grid paramaters.

Utilities.Cache.clean_cache!()

Remove all objects from the cache of created objects.

In most cases, this function should not need to be called, unless you are constructing many grid objects, for example when doing a sweep over grid paramaters.