API

DataLayouts

ClimaCore.DataLayouts — Module

ClimaCore.DataLayouts

Notation:

i,j are horizontal node indices within an element

k is the vertical node index within an element

f is the field index

v is the vertical element index in a stack

h is the element stack index

Data layout is specified by the order in which they appear, e.g. IJKFVH indexes the underlying array as [i,j,k,f,v,h]

ClimaCore.DataLayouts.IF — Type

IF{S, Ni, A} <: DataSlab1D{S, Ni}

Backing DataLayout for 1D spectral element slab data.

Nodal element data (I) are contiguous for each S datatype struct field (F) for a single element slab.

A DataSlab1D view can be returned from other Data1D objects by calling slab(data, idx...).

ClimaCore.DataLayouts.IJF — Type

IJF{S, Nij, A} <: DataSlab2D{S, Nij}

Backing DataLayout for 2D spectral element slab data.

Nodal element data (I,J) are contiguous for each S datatype struct field (F) for a single element slab.

A DataSlab2D view can be returned from other Data2D objects by calling slab(data, idx...).

ClimaCore.DataLayouts.VF — Type

VF{S, A} <: DataColumn{S}

Backing DataLayout for 1D FV column data.

Column level data (V) are contiguous for each S datatype struct field (F).

A DataColumn view can be returned from other Data1DX, Data2DX objects by calling column(data, idx...).

ClimaCore.DataLayouts.IFH — Type

IFH{S, Ni, A} <: Data1D{S, Ni}

Backing DataLayout for 1D spectral element slabs.

Element nodal point (I) data is contiguous for each datatype S struct field (F), for each 1D mesh element (H).

ClimaCore.DataLayouts.IJFH — Type

IJFH{S, Nij, A} <: Data2D{S, Nij}

Backing DataLayout for 2D spectral element slabs.

Element nodal point (I,J) data is contiguous for each datatype S struct field (F), for each 2D mesh element slab (H).

ClimaCore.DataLayouts.VIFH — Type

VIFH{S, Ni, A} <: Data1DX{S, Ni}

Backing DataLayout for 1D spectral element slab + extruded 1D FV column data.

Column levels (V) are contiguous for every element nodal point (I) for each datatype S struct field (F), for each 1D mesh element slab (H).

ClimaCore.DataLayouts.VIJFH — Type

VIJFH{S, Nij, A} <: Data2DX{S, Nij}

Backing DataLayout for 2D spectral element slab + extruded 1D FV column data.

Column levels (V) are contiguous for every element nodal point (I, J) for each S datatype struct field (F), for each 2D mesh element slab (H).

Geometry

Coordinates

ClimaCore.Geometry.AbstractPoint — Type

AbstractPoint

Represents a point in space.

The following types are supported:

XPoint(x)

YPoint(y)

ZPoint(z)

XYPoint(x, y)

XZPoint(x, z)

XYZPoint(x, y, z)

LatLongPoint(lat, long)

LatLongZPoint(lat, long, z)

Cartesian1Point(x1)

Cartesian2Point(x2)

Cartesian3Point(x3)

Cartesian12Point(x1, x2)

Cartesian13Point(x1, x3)

Cartesian123Point(x1, x2, x3)

ClimaCore.Geometry.float_type — Function

float_type(T)

Return the floating point type backing T: T can either be an object or a type.

Domains

Types

ClimaCore.Domains.AbstractDomain — Type

AbstractDomain

A domain represents a region of space.

ClimaCore.Domains.IntervalDomain — Type

IntervalDomain(coord⁻, coord⁺; periodic=true)
IntervalDomain(coord⁻, coord⁺; boundary_names::Tuple{Symbol,Symbol})

Construct a IntervalDomain, the closed interval is given by coord⁻, coord⁺ coordinate arguments.

Either a periodic or boundary_names keyword argument is required.

ClimaCore.Domains.RectangleDomain — Type

RectangleDomain(x1::ClosedInterval, x2::ClosedInterval;
    x1boundary::Tuple{Symbol,Symbol},
    x2boundary::Tuple{Symbol,Symbol},
    x1periodic = false,
    x2periodic = false,
)

Construct a RectangularDomain in the horizontal. If a given x1 or x2 boundary is not periodic, then x1boundary or x2boundary boundary name keyword arguments must be supplied.

ClimaCore.Domains.SphereDomain — Type

SphereDomain(radius)

A domain representing the surface of a sphere with radius radius.

Interfaces

ClimaCore.Domains.boundary_names — Function

boundary_names(obj::Union{AbstractDomain, AbstractMesh, AbstractTopology})

A tuple or vector of unique boundary names of a spatial domain.

Meshes

Mesh types

ClimaCore.Meshes.AbstractMesh — Type

AbstractMesh{dim}

A Mesh is an object which represents how we discretize a domain into elements.

It should be lightweight (i.e. exists on all MPI ranks), e.g for meshes stored in a file, it would contain the filename.

Face and vertex numbering

In 1D, faces and vertices are the same, and both are numbered [1,2].

In 2D, a face is a line segment between to vertices, and both are numbered [1,2,3,4], in a counter-clockwise direction.

 v4        f3        v3
   o-----------------o
   |                 |	    face    vertices
   |                 |	      f1 =>  v1 v2
f4 |                 | f2     f2 =>  v2 v3
   |                 |	      f3 =>  v3 v4
   |                 |        f4 =>  v4 v1
   |                 |
   o-----------------o
  v1       f1        v2

Interface

A subtype of AbstractMesh should define the following methods:

domain(mesh)

elements(mesh)

is_boundary_face(mesh, elem, face)

boundary_face_name(mesh, elem, face)

opposing_face(mesh, elem, face)

coordinates(mesh, elem, vert)

containing_element (optional)

The following types/methods are provided by AbstractMesh:

SharedVertices(mesh, elem, vert)

face_connectivity_matrix(mesh[,elemorder])

vertex_connectivity_matrix(mesh[,elemorder])

ClimaCore.Meshes.IntervalMesh — Type

IntervalMesh <: AbstractMesh

A 1D mesh on an IntervalDomain.

Constuctors

IntervalMesh(domain::IntervalDomain, faces::AbstractVector)

Construct a 1D mesh with face locations at faces.

IntervalMesh(domain::IntervalDomain[, stetching=Uniform()]; nelems=)

Constuct a 1D mesh on domain with nelems elements, using stretching. Possible values of stretching are:

Uniform()

ExponentialStretching(H)

GeneralizedExponentialStretching(dz_surface, dz_top)

ClimaCore.Meshes.RectilinearMesh — Type

RectilinearMesh <: AbstractMesh2D

Constructors

RectilinearMesh(domain::RectangleDomain, n1, n2)

Construct a RectilinearMesh of equally-spaced n1 by n2 elements on domain.

RectilinearMesh(intervalmesh1::IntervalMesh1, intervalmesh2::IntervalMesh2)

Construct the product mesh of intervalmesh1 and intervalmesh2.

ClimaCore.Meshes.AbstractCubedSphere — Type

AbstractCubedSphere <: AbstractMesh2D

This is an abstract type of cubed-sphere meshes on SphereDomains. A cubed-sphere mesh has 6 panels, laid out as follows:

                                          :   Panel 1   :
                            +-------------+-------------+
                            |     +x1     |     +x1     |
                            |             |             |
                            |    Panel    |    Panel    |
                            |+x3   5   -x3|-x2   6   +x2|
                            |     -x2     |     -x3     |
                            |             |             |
                            |     -x1     |     -x1     |
              +-------------+-------------+-------------+
              |     -x2     |     -x2     |
              |             |             |
              |    Panel    |    Panel    |
              |+x1   3   -x1|+x3   4   -x3|
              |     +x3     |     -x1     |
              |             |             |
              |     +x2     |     +x2     |
+-------------+-------------+-------------+
|     +x3     |     +x3     |
|             |             |
|    Panel    |    Panel    |
|-x2   1   +x2|+x1   2   -x1|
|     +x1     |     +x2     |
|             |             |
|     -x3     |     -x3     |
+-------------+-------------+
:   Panel 6   :

This is the same panel ordering used by the S2 Geometry library (though we use 1-based instead of 0-based numering).

Elements are indexed by a CartesianIndex{3} object, where the components are:

horizontal element index (left to right) within each panel.

vertical element index (bottom to top) within each panel.

panel number

Subtypes should have the following fields:

domain: a SphereDomain

ne: number of elements across each panel

External links

S2Geometry library

MIT GCM exch2

ClimaCore.Meshes.EquiangularCubedSphere — Type

EquiangularCubedSphere <: AbstractCubedSphere

An equiangular gnomonic mesh proposed by C. Ronchi, R. Iacono, P. S. Paolucci (1996). Uses the element indexing convention of AbstractCubedSphere.

Constructors

EquiangularCubedSphere(
    domain::Domains.SphereDomain,
    ne::Integer,
    localelementmap=NormalizedBilinearMap()
    )

Constuct an EquiangularCubedSphere on domain with ne elements across each panel.

ClimaCore.Meshes.EquidistantCubedSphere — Type

EquidistantCubedSphere <: AbstractCubedSphere

An equidistant gnomonic mesh outlined in M. Rančić, R. J. Purser, F. Mesinger (1996) and Ramachandran D Nair, Stephen J Thomas, Richard D Loft (2005). Uses the element indexing convention of AbstractCubedSphere.

Constructors

EquidistantCubedSphere(domain::Domains.SphereDomain, ne::Integer)

Constuct an EquidistantCubedSphere on domain with ne elements across each panel.

ClimaCore.Meshes.ConformalCubedSphere — Type

ConformalCubedSphere <: AbstractCubedSphere

A conformal mesh outlined in M. Rančić, R. J. Purser, F. Mesinger (1996). Uses the element indexing convention of AbstractCubedSphere.

Constructors

ConformalCubedSphere(domain::Domains.SphereDomain, ne::Integer)

Constuct an ConformalCubedSphere on domain with ne elements across each panel.

Local element map

ClimaCore.Meshes.LocalElementMap — Type

LocalElementMap

An abstract type of mappings from the reference element to a physical domain.

Missing docstring.

Missing docstring for Meshes.LocalElementMap. Check Documenter's build log for details.

ClimaCore.Meshes.IntrinsicMap — Type

IntrinsicMap()

This LocalElementMap uses the intrinsic mapping of the cubed sphere to map the reference element to the physical domain.

ClimaCore.Meshes.NormalizedBilinearMap — Type

NormalizedBilinearMap()

The LocalElementMap for meshes on spherical domains of O. Guba, M. A. Taylor, P. A. Ullrich, J. R. Overfelt, M. N. Levy (2014). It uses bilinear interpolation between the Cartesian coordinates of the element vertices, then normalizes the result to lie on the sphere.

Mesh stretching

ClimaCore.Meshes.Uniform — Type

Uniform()

Use uniformly-sized elements.

ClimaCore.Meshes.ExponentialStretching — Type

ExponentialStretching(H)

Apply exponential stretching to the domain when constructing elements. H is the scale height (a typical atmospheric scale height H ≈ 7.5km).

For an interval $[z_0,z_1]$, this makes the elements uniformally spaced in $\zeta$, where

\[\zeta = \frac{1 - e^{-\eta/h}}{1-e^{-1/h}},\]

where $\eta = \frac{z - z_0}{z_1-z_0}$, and $h = \frac{H}{z_1-z_0}$ is the non-dimensional scale height.

Interfaces

ClimaCore.Meshes.domain — Function

Meshes.domain(mesh::AbstractMesh)

The Domains.AbstractDomain on which the mesh is defined.

ClimaCore.Meshes.elements — Function

Meshes.elements(mesh::AbstractMesh)

An iterator over the elements of a mesh. Elements of a mesh can be of any type.

ClimaCore.Meshes.nelements — Function

nelements(mesh::AbstractMesh)

The number of elements in the mesh.

ClimaCore.Meshes.is_boundary_face — Function

Meshes.is_boundary_face(mesh::AbstractMesh, elem, face::Int)::Bool

Determine whether face face of element elem is on the boundary of mesh.

elem should be an element of elements(mesh).

ClimaCore.Meshes.boundary_face_name — Function

Meshes.boundary_face_name(mesh::AbstractMesh, elem, face::Int)::Union{Symbol,Nothing}

The name of the boundary facing face of element elem, or nothing if it is not on the boundary.

ClimaCore.Meshes.opposing_face — Function

opelem, opface, reversed = Meshes.opposing_face(mesh::AbstractMesh, elem, face::Int)

The element and face (opelem, opface) that oppose face face of element elem.

ClimaCore.Meshes.coordinates — Function

Meshes.coordinates(mesh, elem, vert::Int)
Meshes.coordinates(mesh, elem, ξ::SVector)

Return the physical coordinates of a point in an element elem of mesh. The position of the point can either be a vertex number vert or the coordinates ξ in the reference element.

ClimaCore.Meshes.containing_element — Function

elem = Meshes.containing_element(mesh::AbstractMesh, coord)

The element elem in mesh containing the coordinate coord. If the coordinate falls on the boundary between two or more elements, an arbitrary element is chosen.

ClimaCore.Meshes.reference_coordinates — Function

ξ = Meshes.reference_coordinates(mesh::AbstractMesh, elem, coord)

An SVector of coordinates in the reference element such that

Meshes.coordinates(mesh, elem, ξ) == coord

This can be used for interpolation to a specific point.

ClimaCore.Meshes.SharedVertices — Type

Meshes.SharedVertices(mesh, elem, vert)

An iterator over (element, vertex) pairs that are shared with (elem,vert).

ClimaCore.Meshes.face_connectivity_matrix — Function

M = Meshes.face_connectivity_matrix(mesh, elemorder = elements(mesh))

Construct a Bool-valued SparseCSCMatrix containing the face connections of mesh. Elements are indexed according to elemorder.

Note that M[i,i] == true only if two distinct faces of element i are connected.

ClimaCore.Meshes.vertex_connectivity_matrix — Function

M = Meshes.vertex_connectivity_matrix(mesh, elemorder = elements(mesh))

Construct a Bool-valued SparseCSCMatrix containing the vertex connections of mesh. Elements are indexed according to elemorder.

Note that M[i,i] == true only if two distinct vertices of element i are connected.

ClimaCore.Meshes.linearindices — Function

Meshes.linearindices(elemorder)

Given a data structure elemorder[i] = elem that orders elements, construct the inverse map from orderindex = linearindices(elemorder) such that orderindex[elem] = i.

This will try to use the most efficient structure available.

Topologies

Types

ClimaCore.Topologies.AbstractTopology — Type

AbstractTopology

Subtypes of AbstractHorizontalTopology define connectiveness of a mesh in the horizontal domain.

Interfaces

`nelems'

nlocalelems

nneighbors

nsendelems

nghostelems

localelemindex

vertex_coordinates

opposing_face

face_node_index

interior_faces

ghost_faces

vertex_node_index

vertices

neighboring_elements

interior_vertices

ghost_vertices

neighbors

boundary_tags

boundary_tag

boundary_faces

ClimaCore.Topologies.IntervalTopology — Type

IntervalTopology(mesh::IntervalMesh)

A sequential topology on an Meshes.IntervalMesh.

ClimaCore.Topologies.Topology2D — Type

Topology2D(mesh::AbstractMesh2D, elemorder=Mesh.elements(mesh))

This is a generic non-distributed topology for 2D meshes. elemorder is a vector or other linear ordering of the Mesh.elements(mesh).

Interfaces

ClimaCore.Topologies.nlocalelems — Function

nlocalelems(topology)

The number of local elements in topology.

ClimaCore.Topologies.vertex_coordinates — Function

(c1,c2,c3,c4) = vertex_coordinates(topology, elem)

The coordinates of the 4 vertices of element elem.

ClimaCore.Topologies.opposing_face — Function

(opelem, opface, reversed) = opposing_face(topology, elem, face)

The opposing face of face number face of element elem in topology.

opelem is the opposing element number, 0 for a boundary, negative for a ghost element

opface is the opposite face number, or boundary face number if a boundary

reversed indicates whether the opposing face has the opposite orientation.

ClimaCore.Topologies.interior_faces — Function

interior_faces(topology::AbstractTopology)

An iterator over the interior faces of topology. Each element of the iterator is a 5-tuple the form

(elem1, face1, elem2, face2, reversed)

where elemX, faceX are the element and face numbers, and reversed indicates whether they have opposing orientations.

ClimaCore.Topologies.boundary_tags — Function

boundary_tags(topology)

A Tuple or NamedTuple of the boundary tags of the topology. A boundary tag is an integer that uniquely identifies a boundary.

ClimaCore.Topologies.boundary_tag — Function

boundary_tag(topology, name::Symbol)

The boundary tag of the topology for boundary name name. A boundary tag is an integer that uniquely identifies a boundary.

ClimaCore.Topologies.boundary_faces — Function

boundary_faces(topology, boundarytag)

An iterator over the faces of topology which face the boundary with tag boundarytag. Each element of the iterator is an (elem, face) pair.

ClimaCore.Topologies.vertices — Function

vertices(topology)

An iterator over the unique (shared) vertices of the topology topology. Each vertex is an iterator over (element, vertex_number) pairs.

Spaces

ClimaCore.Spaces — Module

Meshes

domain

topology

coordinates

metric terms (inverse partial derivatives)

quadrature rules and weights

References / notes

ceed

QA

Quadratures

ClimaCore.Spaces.Quadratures.QuadratureStyle — Type

QuadratureStyle

Quadrature style supertype. See sub-types:

GLL

GL

Uniform

ClimaCore.Spaces.Quadratures.GLL — Type

GLL{Nq}()

Gauss-Legendre-Lobatto quadrature using Nq quadrature points.

ClimaCore.Spaces.Quadratures.GL — Type

GL{Nq}()

Gauss-Legendre quadrature using Nq quadrature points.

ClimaCore.Spaces.Quadratures.Uniform — Type

Uniform{Nq}()

Uniformly-spaced quadrature.

ClimaCore.Spaces.Quadratures.degrees_of_freedom — Function

degrees_of_freedom(QuadratureStyle) -> Int

Returns the degreesoffreedom of the QuadratureStyle concrete type

ClimaCore.Spaces.Quadratures.polynomial_degree — Function

polynomial_degree(QuadratureStyle) -> Int

Returns the polynomial degree of the QuadratureStyle concrete type

ClimaCore.Spaces.Quadratures.quadrature_points — Function

points, weights = quadrature_points(::Type{FT}, quadrature_style)

The points and weights of the quadrature rule in floating point type FT.

ClimaCore.Spaces.Quadratures.barycentric_weights — Function

barycentric_weights(x::SVector{Nq}) where {Nq}

The barycentric weights associated with the array of point locations x:

\[w_j = \frac{1}{\prod_{k \ne j} (x_i - x_j)}\]

See Jean-Paul Berrut, Lloyd N Trefethen (2004), equation 3.2.

ClimaCore.Spaces.Quadratures.interpolation_matrix — Function

interpolation_matrix(x::SVector, r::SVector{Nq})

The matrix which interpolates the Lagrange polynomial of degree Nq-1 through the points r, to points x. The matrix coefficients are computed using the Barycentric formula of Jean-Paul Berrut, Lloyd N Trefethen (2004), section 4:

\[I_{ij} = \begin{cases} 1 & \text{if } x_i = r_j, \\ 0 & \text{if } x_i = r_k \text{ for } k \ne j, \\ \frac{\displaystyle \frac{w_j}{x_i - r_j}}{\displaystyle \sum_k \frac{w_k}{x_i - r_k}} & \text{otherwise,} \end{cases}\]

where $w_j$ are the barycentric weights, see barycentric_weights.

ClimaCore.Spaces.Quadratures.differentiation_matrix — Function

differentiation_matrix(r::SVector{Nq, T}) where {Nq, T}

The spectral differentiation matrix for the Lagrange polynomial of degree Nq-1 interpolating at points r.

The matrix coefficients are computed using the Jean-Paul Berrut, Lloyd N Trefethen (2004), section 9.3:

\[D_{ij} = \begin{cases} \displaystyle \frac{w_j}{w_i (x_i - x_j)} &\text{ if } i \ne j \\ -\sum_{k \ne j} D_{kj} &\text{ if } i = j \end{cases}\]

where $w_j$ are the barycentric weights, see barycentric_weights.

differentiation_matrix(FT, quadstyle::QuadratureStyle)

The spectral differentiation matrix at the quadrature points of quadstyle, using floating point types FT.

ClimaCore.Spaces.Quadratures.orthonormal_poly — Function

V = orthonormal_poly(points, quad)

V_{ij} contains the j-1th Legendre polynomial evaluated at points[i]. i.e. it is the mapping from the modal to the nodal representation.

RecursiveApply

ClimaCore.RecursiveApply — Module

RecursiveApply

This module contains operators to recurse over nested Tuples or NamedTuples.

To extend to another type T, define RecursiveApply.rmap(fn, args::T...)

ClimaCore.RecursiveApply.tuplemap — Function

tuplemap(fn::Function, tup)

A map impl for mapping function fn a tuple argument tup

Currently just calls Base.map behind the scenes but is left stubbed out for potential specialization in the future.

tuplemap(fn::Function, tup1, tup2)

A map impl for mapping function fn over tup1, tup2 tuple arguments.

Currently just calls Base.map behind the scenes but is left stubbed out for potential specialization in the future.

Fields

ClimaCore.Fields.Field — Type

Field(values, space)

A set of values defined at each point of a space.

ClimaCore.Fields.coordinate_field — Function

coordinate_field(space::AbstractSpace)

Construct a Field of the coordinates of the space.

ClimaCore.Fields.local_geometry_field — Function

local_geometry_field(space::AbstractSpace)

Construct a Field of the LocalGeometry of the space.

Base.zeros — Method

zeros(space::AbstractSpace)

Construct a field on space that is zero everywhere.

Base.ones — Method

ones(space::AbstractSpace)

Construct a field on space that is one everywhere.

Base.sum — Method

sum([f=identity,]v::Field)

Approximate integration of v or f.(v) over the domain. In an AbstractSpectralElementSpace, an integral over the entire space is computed by summation over the elements of the integrand multiplied by the Jacobian determinants and the quadrature weights at each node within an element. Hence, sum is computed by summation of the field values multiplied by the Jacobian determinants and quadrature weights:

\[\sum_i f(v_i) W_i J_i \approx \int_\Omega f(v) \, d \Omega\]

where $v_i$ is the value at each node, and $f$ is the identity function if not specified.

Statistics.mean — Method

mean([f=identity, ]v::Field)

The mean value of field or f.(field) over the domain, weighted by area. Similar to sum, in an AbstractSpectralElementSpace, this is computed by summation of the field values multiplied by the Jacobian determinants and quadrature weights:

\[\frac{\sum_i f(v_i) W_i J_i}{\sum_i W_i J_i} \approx \frac{\int_\Omega f(v) \, d \Omega}{\int_\Omega \, d \Omega}\]

where $v_i$ is the Field value at each node, and $f$ is the identity function if not specified.

LinearAlgebra.norm — Method

norm(v::Field, p=2; normalize=true)

The approximate $L^p$ norm of v, where $L^p$ represents the space of measurable functions for which the p-th power of the absolute value is Lebesgue integrable, that is:

\[\| v \|_p = \left( \int_\Omega |v|^p d \Omega \right)^{1/p}\]

where $|v|$ is defined to be the absolute value if `v is a scalar-valued Field, or the 2-norm if it is a vector-valued Field or composite Field (see LinearAlgebra.norm). Similar to sum and mean, in an AbstractSpectralElementSpace, this is computed by summation of the field values multiplied by the Jacobian determinants and quadrature weights. If normalize=true (the default), then internally the discrete norm is divided by the sum of the Jacobian determinants and quadrature weights:

\[\left(\frac{\sum_i |v_i|^p W_i J_i}{\sum_i W_i J_i}\right)^{1/p} \approx \left(\frac{\int_\Omega |v|^p \, d \Omega}{\int_\Omega \, d \Omega}\right)^{1/p}\]

If p=Inf, then the norm is the maximum of the absolute values

\[\max_i |v_i| \approx \sup_{\Omega} |v|\]

Consequently all norms should have the same units for all $p$ (being the same as calling norm on a single value).

If normalize=false, then the denominator term is omitted, and so the result will be the norm as described above multiplied by the length/area/volume of $\Omega$ to the power of $1/p$.

Limiters

Interfaces

ClimaCore.Limiters.quasimonotone_limiter! — Function

quasimonotone_limiter!(ρq, ρ, min_ρq, max_ρq; rtol)

Arguments:

ρq: tracer density Field, where q denotes tracer concentration per unit mass

ρ: fluid density Field

min_ρq: Array of min(ρq) per element

max_ρq: Array of max(ρq) per element

rtol: relative tolerance needed to solve element-wise optimization problem

This limiter is inspired by the one presented in Guba et al Oksana Guba, Mark Taylor, Amik St-Cyr (2014). In the reference paper, it is denoted by OP1, and is outlined in eqs. (37)-(40). Quasimonotone here is meant to be monotone with respect to the spectral element nodal values. This limiter involves solving a constrained optimization problem (a weighted least square problem up to a fixed tolerance denoted by rtol) that is completely local to each element. As in HOMME, the implementation idea here is the following: we need to find a grid field which is closest to the initial field (in terms of weighted sum), but satisfies the min/max constraints. So, first we find values that do not satisfy constraints and bring these values to a closest constraint. This way we introduce some mass change (mass_change), which we then redistribute so that the l2 error is smallest. This redistribution might violate constraints; thus, we do a few iterations (typically a couple).