API

DataLayouts

ClimaCore.DataLayouts — Module

ClimaCore.DataLayouts

Notation:

i,j are horizontal node indices within an element

k is the vertical node index within an element

f is the field index

v is the vertical element index in a stack

h is the element stack index

Data layout is specified by the order in which they appear, e.g. IJKFVH indexes the underlying array as [i,j,k,f,v,h]

ClimaCore.DataLayouts.IJFH — Type

IJFH{S,Nij}(ArrayType, nelements)

Construct an IJFH structure given the backing ArrayType, quadrature degrees of freedom Nij, and the number of mesh elements nelements.

Meshes

ClimaCore.Meshes.EquispacedRectangleMesh — Type

EquispacedRectangleMesh(domain::RectangleDomain, n1::Integer, n2::Integer)

A regular AbstractMesh of domain with n1 elements in dimension 1, and n2 in dimension 2.

Geometry

Missing docstring.

Missing docstring for Geometry.Cartesian12Vector. Check Documenter's build log for details.

Topologies

ClimaCore.Topologies — Module

ClimaCore.Topologies

Objects describing the horizontal connections between elements.

All elements are quadrilaterals, using the face and vertex numbering convention from p4est:

          4
      3-------4
 ^    |       |
 |  1 |       | 2
x2    |       |
      1-------2
          3
        x1-->

Spaces

ClimaCore.Spaces — Module

Meshes

domain

topology

coordinates

metric terms (inverse partial derivatives)

quadrature rules and weights

References / notes

ceed

QA

ClimaCore.Spaces.Quadratures.degrees_of_freedom — Function

degrees_of_freedom(QuadratureStyle) -> Int

Returns the degreesoffreedom of the QuadratureStyle concrete type

ClimaCore.Spaces.Quadratures.GL — Type

GL{Nq}()

Gauss-Legendre quadrature using Nq quadrature points.

ClimaCore.Spaces.Quadratures.Uniform — Type

Uniform{Nq}()

Uniformly-spaced quadrature.

ClimaCore.Spaces.Quadratures.quadrature_points — Function

points, weights = quadrature_points(::Type{FT}, quadrature_style)

The points and weights of the quadrature rule in floating point type FT.

ClimaCore.Spaces.Quadratures.GLL — Type

GLL{Nq}()

Gauss-Legendre-Lobatto quadrature using Nq quadrature points.

ClimaCore.Spaces.Quadratures.polynomial_degree — Function

polynomial_degree(QuadratureStyle) -> Int

Returns the polynomial degree of the QuadratureStyle concrete type

ClimaCore.Spaces.Quadratures.QuadratureStyle — Type

QuadratureStyle

Quadrature style supertype. See sub-types:

GLL

GL

Uniform

RecursiveApply

ClimaCore.RecursiveApply — Module

RecursiveApply

This module contains operators to recurse over nested Tuples or NamedTuples.

To extend to another type T, define RecursiveApply.rmap(fn, args::T...)