API

DataLayouts

ClimaCore.DataLayouts — Module

ClimaCore.DataLayouts

Notation:

i,j are horizontal node indices within an element

k is the vertical node index within an element

f is the field index

v is the vertical element index in a stack

h is the element stack index

Data layout is specified by the order in which they appear, e.g. IJKFVH indexes the underlying array as [i,j,k,f,v,h]

ClimaCore.DataLayouts.IJFH — Type

IJFH{S, Nij, A} <: Data2D{S, Nij}

Backing DataLayout for 2D spectral element slabs.

Element nodal point (I,J) data is contiguous for each datatype S struct field (F), for each 2D mesh element slab (H).

Geometry

Missing docstring.

Missing docstring for Geometry.Cartesian12Vector. Check Documenter's build log for details.

Domains

ClimaCore.Domains.IntervalDomain — Type

IntervalDomain(coord⁻, coord⁺; periodic=true)
IntervalDomain(coord⁻, coord⁺; boundary_tags::Tuple{Symbol,Symbol})

Construct a IntervalDomain, the closed interval is given by coord⁻, coord⁺ coordinate arguments.

Either a periodic or boundary_tags keyword argument is required.

IntervalDomain(coords::ClosedInterval; boundary_tags::Tuple{Symbol,Symbol})

Construct a IntervalDomain, over the closed coordinate interval coords Because IntervalDomain does not support periodic boundary conditions, the boundary_tags keyword arugment must be supplied.

ClimaCore.Domains.RectangleDomain — Type

RectangleDomain(x1::ClosedInterval, x2::ClosedInterval;
    x1boundary::Tuple{Symbol,Symbol},
    x2boundary::Tuple{Symbol,Symbol},
    x1periodic = false,
    x2periodic = false,
)

Construct a RectangularDomain in the horizontal. If a given x1 or x2 boundary is not periodic, then x1boundary or x2boundary boundary tag keyword arguments must be supplied.

Meshes

ClimaCore.Meshes.IntervalMesh — Type

IntervalMesh <: AbstractMesh

A 1D mesh on an IntervalDomain.

Constuctors

IntervalMesh(domain::IntervalDomain, faces::AbstractVector)

Construct a 1D mesh with face locations at faces.

IntervalMesh(domain::IntervalDomain[, stetching=Uniform()]; nelems=)

Constuct a 1D mesh on domain with nelems elements, using stretching. Possible values of stretching are:

Uniform()

ExponentialStretching(H)

ClimaCore.Meshes.Uniform — Type

Uniform()

Use uniformly-sized elements.

ClimaCore.Meshes.ExponentialStretching — Type

ExponentialStretching(H)

Apply exponential stretching to the domain when constructing elements. H is the scale height (a typical atmospheric scale height H ≈ 7.5e3km).

For an interval $[z_0,z_1]$, this makes the elements uniformally spaced in $\zeta$, where

\[\zeta = \frac{1 - e^{-\eta/h}}{1-e^{-1/h}},\]

where $\eta = \frac{z - z_0}{z_1-z_0}$, and $h = \frac{H}{z_1-z_0}$ is the non-dimensional scale height.

ClimaCore.Meshes.EquispacedRectangleMesh — Type

EquispacedRectangleMesh(domain::RectangleDomain, n1::Integer, n2::Integer)

A regular AbstractMesh of domain with n1 elements in dimension 1, and n2 in dimension 2.

Topologies

ClimaCore.Topologies — Module

ClimaCore.Topologies

Objects describing the horizontal connections between elements.

All elements are quadrilaterals, using the following face and vertex numbering convention:

          3
      4-------3
 ^    |       |
 |  4 |       | 2
ξ2    |       |
      1-------2
          1
        ξ1-->

ClimaCore.Topologies.IntervalTopology — Type

IntervalTopology(mesh::IntervalMesh)

A sequential topology on an Meshes.IntervalMesh.

Spaces

ClimaCore.Spaces — Module

Meshes

domain

topology

coordinates

metric terms (inverse partial derivatives)

quadrature rules and weights

References / notes

ceed

QA

ClimaCore.Spaces.Quadratures.degrees_of_freedom — Function

degrees_of_freedom(QuadratureStyle) -> Int

Returns the degreesoffreedom of the QuadratureStyle concrete type

ClimaCore.Spaces.Quadratures.GL — Type

GL{Nq}()

Gauss-Legendre quadrature using Nq quadrature points.

ClimaCore.Spaces.Quadratures.Uniform — Type

Uniform{Nq}()

Uniformly-spaced quadrature.

ClimaCore.Spaces.Quadratures.quadrature_points — Function

points, weights = quadrature_points(::Type{FT}, quadrature_style)

The points and weights of the quadrature rule in floating point type FT.

ClimaCore.Spaces.Quadratures.GLL — Type

GLL{Nq}()

Gauss-Legendre-Lobatto quadrature using Nq quadrature points.

ClimaCore.Spaces.Quadratures.polynomial_degree — Function

polynomial_degree(QuadratureStyle) -> Int

Returns the polynomial degree of the QuadratureStyle concrete type

ClimaCore.Spaces.Quadratures.QuadratureStyle — Type

QuadratureStyle

Quadrature style supertype. See sub-types:

GLL

GL

Uniform

RecursiveApply

ClimaCore.RecursiveApply — Module

RecursiveApply

This module contains operators to recurse over nested Tuples or NamedTuples.

To extend to another type T, define RecursiveApply.rmap(fn, args::T...)