ClimaCoreMakie.jl

ClimaCoreMakie.jl provides functionality for plotting ClimaCore fields extending the Makie.jl package.

ClimaCoreMakie.fieldheatmap! — Function

fieldheatmap(field)

Plots a heatmap of a field.

Attributes

Inherited from Makie.mesh.

colormap::Union{Symbol, Vector{<:Colorant}} = :viridis` sets the colormap that is sampled for numeric colors.

colorrange::Tuple{<:Real, <:Real} sets the values representing the start and end points of colormap.

nan_color::Union{Symbol, <:Colorant} = RGBAf(0,0,0,0) sets a replacement color for color = NaN.

lowclip::Union{Automatic, Symbol, <:Colorant} = automatic sets a color for any value below the colorrange.

highclip::Union{Automatic, Symbol, <:Colorant} = automatic sets a color for any value above the colorrange.

ClimaCoreMakie.fieldcontourf! — Function

fieldcontourf(field::Field)

Plots a 2D filled contour plot of a field.

Attributes

These are inherited from Makie.tricontourf:

levels = 10 can be either an Int which results in n bands delimited by n+1 equally spaced levels, or it can be an AbstractVector{<:Real} that lists n consecutive edges from low to high, which result in n-1 bands.

mode = :normal sets the way in which a vector of levels is interpreted, if it's set to :relative, each number is interpreted as a fraction between the minimum and maximum values of zs. For example, levels = 0.1:0.1:1.0 would exclude the lower 10% of data.

extendlow = nothing. This sets the color of an optional additional band from minimum(zs) to the lowest value in levels. If it's :auto, the lower end of the colormap is picked and the remaining colors are shifted accordingly. If it's any color representation, this color is used. If it's nothing, no band is added.

extendhigh = nothing. This sets the color of an optional additional band from the highest value of levels to maximum(zs). If it's :auto, the high end of the colormap is picked and the remaining colors are shifted accordingly. If it's any color representation, this color is used. If it's nothing, no band is added.

color sets the color of the plot. It can be given as a named color Symbol or a Colors.Colorant. Transparency can be included either directly as an alpha value in the Colorant or as an additional float in a tuple (color, alpha). The color can also be set for each scattered marker by passing a Vector of colors or be used to index the colormap by passing a Real number or Vector{<: Real}.

colormap::Union{Symbol, Vector{<:Colorant}} = :viridis sets the colormap from which the band colors are sampled.