Operators

Finite difference operators

Interpolation operators

ClimateMachineCore.Operators.InterpolateC2F — Type

InterpolateC2F(;boundaries..)

Interpolate from center to face. Supported boundary conditions are:

SetValue(val): set the value at the boundary face to be val.

SetGradient: set the value at the boundary such that the gradient is val.

Extrapolate: use the closest interior point as the boundary value

ClimateMachineCore.Operators.InterpolateF2C — Type

InterpolateF2C()

Interpolate from face to center mesh. No boundary conditions are required (or supported).

ClimateMachineCore.Operators.UpwindLeftC2F — Type

UpwindLeftC2F(;boundaries)

Interpolate from the left. Only the left boundary condition should be set:

SetValue(val): set the value to be val on the boundary.

ClimateMachineCore.Operators.SetBoundaryOperator — Type

SetBoundaryOperator(;boundaries...)

This operator only modifies the values at the boundary:

SetValue(val): set the value to be val on the boundary.

Gradient operators

ClimateMachineCore.Operators.GradientF2C — Type

GradientF2C(;boundaryname=boundarycondition...)

Centered-difference gradient operator from a FaceFiniteDifferenceSpace to a CenterFiniteDifferenceSpace, applying the relevant boundary conditions. These can be:

by default, the current value at the boundary face will be used.

SetValue(val): calculate the gradient assuming the value at the boundary is val.

Extrapolate(): set the value at the center closest to the boundary to be the same as the neighbouring interior value.

ClimateMachineCore.Operators.GradientC2F — Type

GradientC2F(;boundaries)

Centered-difference gradient operator from a CenterFiniteDifferenceSpace to a FaceFiniteDifferenceSpace, applying the relevant boundary conditions. These can be:

SetValue(val): calculate the gradient assuming the value at the boundary is val.

SetGradient(val): set the value of the gradient at the boundary to be val.

Boundary conditions

ClimateMachineCore.Operators.SetValue — Type

SetValue(val)

Set the value at the boundary to be val. In the case of gradient operators, this will set the input value from which the gradient is computed.

ClimateMachineCore.Operators.SetGradient — Type

SetGradient(val)

Set the gradient at the boundary to be val. In the case of gradient operators this will set the output value of the gradient.

ClimateMachineCore.Operators.Extrapolate — Type

Extrapolate()

Set the value at the boundary to be the same as the closest interior point.