Operators

Finite difference operators

Interpolation operators

ClimaCore.Operators.InterpolateC2F — Type

I = InterpolateC2F(;boundaries..)
I.(x)

Interpolate a center-valued field x to faces, using the stencil

\[I(x)[i] = \frac{1}{2} (x[i+\tfrac{1}{2}] + x[i-\tfrac{1}{2}])\]

Supported boundary conditions are:

SetValue(x₀): set the value at the boundary face to be x₀. On the left boundary the stencil is

\[I(x)[\tfrac{1}{2}] = x₀\]

SetGradient(v): set the value at the boundary such that the gradient is v. At the left boundary the stencil is

\[I(x)[\tfrac{1}{2}] = x[1] - \frac{1}{2} v³\]

Extrapolate: use the closest interior point as the boundary value. At the left boundary the stencil is

\[I(x)[\tfrac{1}{2}] = x[1]\]

ClimaCore.Operators.InterpolateF2C — Type

InterpolateF2C()

Interpolate from face to center mesh. No boundary conditions are required (or supported).

ClimaCore.Operators.UpwindBiasedProductC2F — Type

U = UpwindBiasedProductC2F(;boundaries)
U.(v, x)

Compute the product of a face-valued vector field v and a center-valued field x at cell faces by upwinding x according to v

\[U(v,x)[i] = \begin{cases} v[i] x[i-\tfrac{1}{2}] \quad v^3[i] > 0 \\ v[i] x[i+\tfrac{1}{2}] \quad v^3[i] < 0 \end{cases}\]

Supported boundary conditions are:

SetValue(x₀): set the value of x to be x₀ on the boundary. On the left boundary the stencil is
\[U(v,x)[\tfrac{1}{2}] = \begin{cases} v[\tfrac{1}{2}] x_0 \quad v^3[\tfrac{1}{2}] > 0 \\ v[\tfrac{1}{2}] x[1] \quad v^3[\tfrac{1}{2}] < 0 \end{cases}\]

Extrapolate(): set the value of x to be the same as the closest interior point. On the left boundary, the stencil is U(v,x)[\tfrac{1}{2}] = U(v,x)[1 + \tfrac{1}{2}]

ClimaCore.Operators.LeftBiasedC2F — Type

L = LeftBiasedC2F(;boundaries)
L.(x)

Interpolate a left-value field to a face-valued field from the left.

\[L(x)[i] = x[i-\tfrac{1}{2}]\]

Only the left boundary condition should be set. Currently supported is:

SetValue(x₀): set the value to be x₀ on the boundary.

\[L(x)[\tfrac{1}{2}] = x_0\]

ClimaCore.Operators.RightBiasedC2F — Type

R = RightBiasedC2F(;boundaries)
R.(x)

Interpolate a center-valued field to a face-valued field from the right.

\[R(x)[i] = x[i+\tfrac{1}{2}]\]

Only the right boundary condition should be set. Currently supported is:

SetValue(x₀): set the value to be x₀ on the boundary.

\[R(x)[n+\tfrac{1}{2}] = x_0\]

Advection operators

ClimaCore.Operators.AdvectionF2F — Type

A = AdvectionF2F(;boundaries)
A.(v, θ)

Vertical advection operator at cell faces, for a face-valued velocity field v and face-valued variables θ, approximating $v^3 \partial_3 \theta$.

It uses the following stencil

\[A(v,θ)[i] = \frac{1}{2} (θ[i+1] - θ[i-1]) v³[i]\]

No boundary conditions are currently supported. ```

ClimaCore.Operators.AdvectionC2C — Type

A = AdvectionC2C(;boundaries)
A.(v, θ)

Vertical advection operator at cell centers, for cell face velocity field v cell center variables θ, approximating $v^3 \partial_3 \theta$.

It uses the following stencil

\[A(v,θ)[i] = \frac{1}{2} \{ (θ[i+1] - θ[i]) v³[i+\tfrac{1}{2}] + (θ[i] - θ[i-1])v³[i-\tfrac{1}{2}]\}\]

Supported boundary conditions:

SetValue(θ₀): set the value of θ at the boundary face to be θ₀. At the lower boundary, this is:

\[A(v,θ)[1] = \frac{1}{2} \{ (θ[2] - θ[1]) v³[1 + \tfrac{1}{2}] + (θ[1] - θ₀)v³[\tfrac{1}{2}]\}\]

Extrapolate: use the closest interior point as the boundary value. At the lower boundary, this is:

\[A(v,θ)[1] = (θ[2] - θ[1]) v³[1 + \tfrac{1}{2}] \}\]

Gradient operators

ClimaCore.Operators.GradientF2C — Type

G = GradientF2C(;boundaryname=boundarycondition...)
G.(x)

Compute the gradient of a face-valued field x, returning a center-valued Covariant3 vector field, using the stencil:

\[G(x)[i]^3 = x[i+\tfrac{1}{2}] - x[i-\tfrac{1}{2}]\]

The following boundary conditions are supported:

by default, the value of x at the boundary face will be used.

SetValue(x₀): calculate the gradient assuming the value at the boundary is x₀. For the left boundary, this becomes:

\[G(x)[1]³ = x[1+\tfrac{1}{2}] - x₀\]

Extrapolate(): set the value at the center closest to the boundary

to be the same as the neighbouring interior value. For the left boundary, this becomes:

\[G(x)[1]³ = G(x)[2]³\]

ClimaCore.Operators.GradientC2F — Type

G = GradientC2F(;boundaryname=boundarycondition...)
G.(x)

Compute the gradient of a center-valued field x, returning a face-valued Covariant3 vector field, using the stencil:

\[G(x)[i]^3 = x[i+\tfrac{1}{2}] - x[i-\tfrac{1}{2}]\]

The following boundary conditions are supported:

SetValue(x₀): calculate the gradient assuming the value at the boundary is x₀. For the left boundary, this becomes:
\[G(x)[\tfrac{1}{2}]³ = 2 (x[1] - x₀)\]

SetGradient(v₀): set the value of the gradient at the boundary to be v₀. For the left boundary, this becomes:
\[G(x)[\tfrac{1}{2}] = v₀\]

Divergence operators

ClimaCore.Operators.DivergenceF2C — Type

D = DivergenceF2C(;boundaryname=boundarycondition...)
D.(v)

Compute the vertical contribution to the divergence of a face-valued field vector v, returning a center-valued scalar field, using the stencil

\[D(v)[i] = (Jv³[i+\tfrac{1}{2}] - Jv³[i-\tfrac{1}{2}]) / J[i]\]

where Jv³ is the Jacobian multiplied by the third contravariant component of v.

The following boundary conditions are supported:

by default, the value of v at the boundary face will be used.

SetValue(v₀): calculate the divergence assuming the value at the boundary is v₀. For the left boundary, this becomes:

\[D(v)[1] = (Jv³[1+\tfrac{1}{2}] - Jv³₀) / J[i]\]

Extrapolate(): set the value at the center closest to the boundary to be the same as the neighbouring interior value. For the left boundary, this becomes:

\[D(v)[1]³ = D(v)[2]³\]

ClimaCore.Operators.DivergenceC2F — Type

D = DivergenceC2F(;boundaryname=boundarycondition...)
D.(v)

Compute the vertical contribution to the divergence of a center-valued field vector v, returning a face-valued scalar field, using the stencil

\[D(v)[i] = (Jv³[i+\tfrac{1}{2}] - Jv³[i-\tfrac{1}{2}]) / J[i]\]

where Jv³ is the Jacobian multiplied by the third contravariant component of v.

The following boundary conditions are supported:

SetValue(v₀): calculate the divergence assuming the value at the boundary is v₀. For the left boundary, this becomes:
\[D(v)[\tfrac{1}{2}] = \frac{1}{2} (Jv³[1] - Jv³₀) / J[i]\]

SetDivergence(x): set the value of the divergence at the boundary to be x.
\[D(v)[\tfrac{1}{2}] = x\]

Other

ClimaCore.Operators.SetBoundaryOperator — Type

SetBoundaryOperator(;boundaries...)

This operator only modifies the values at the boundary:

SetValue(val): set the value to be val on the boundary.

Finite difference boundary conditions

ClimaCore.Operators.SetValue — Type

SetValue(val)

Set the value at the boundary to be val. In the case of gradient operators, this will set the input value from which the gradient is computed.

ClimaCore.Operators.SetGradient — Type

SetGradient(val)

Set the gradient at the boundary to be val. In the case of gradient operators this will set the output value of the gradient.

ClimaCore.Operators.SetDivergence — Type

SetDivergence(val)

Set the divergence at the boundary to be val.

ClimaCore.Operators.Extrapolate — Type

Extrapolate()

Set the value at the boundary to be the same as the closest interior point.